Update to version 20190912
The source filename is `jotun_12_09_2019_32415.sh`. This appears to be
in the form `jotun_dd_mm_yyyy_timestamp.sh`, and an unfortunate
side-effect of this issue is that newer package versions don't play well
with `vercmp`'s version comparison logic. Consequently, this PKGBUILD
does some work to separate the date-based version (20190912) from the
filename.