Make all tree-sitter parsers required dependencies
Upstream considers all of these parsers to be required dependencies. The
only one included here that is _strictly_ required today (in the sense
that there is broken functionality without it) is tree-sitter-markdown,
which is required for the LSP hover window. However, since upstream
considers all of the included parsers as dependencies, Arch should
follow suit.